fix: tirei chaves compostas e inclui uniqueConstraint. Exclui classe expertise_qualifier e inclui qualifiers como lista na classe expertise
