+++ src/main/java/com/fatec/springAPI4/entity/Expertise.java
@@ -23,6 +23,9 @@ public class Expertise {
     private String name;
     
     @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "ex_tk_id")
+    @JoinColumn(name = "tk_id")
     private Track track;
+
+    @Column(name = "ex_qualifiers")
+    private List<Qualifier> qualifiers;
 }
