+++ src/main/java/com/fatec/springAPI4/entity/PartnerExpertise.java
@@ -9,15 +9,26 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import javax.persistence.UniqueConstraint;
 
 import java.time.LocalDate;
 
 @Getter @Setter @AllArgsConstructor
 @Entity
-@Table
-public class PartnerQualifier {
-    @EmbeddedId
-    private PartnerExpertisePrimaryData partnerExpertisePrimaryData;
+@Table(uniqueConstraints = { @UniqueConstraint(columnNames = { "pt_id", "ex_id" }) })
+public class PartnerExpertise {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = pt_ex_id)
+    private Long id;
+    
+    @ManyToOne
+    @JoinColumn(name = "pt_id", referencedColumnName = "pt_id")
+    private Partner partnerId;
+    
+    @ManyToOne
+    @JoinColumn(name = "ex_id", referencedColumnName = "ex_id")
+    private Expertise expertiseId;
     
     @Column(name = pt_ex_insert_date)
     private LocalDate insertDate;
